django_1_5/staff-gaia/requirements/base.txt
django>=1.5,<1.5.99
django-grappelli>=2.4,<=2.4.99
django-filebrowser>=3.5,<=3.5.99

django-admin-honeypot
django-compressor==1.4
django-mptt==0.7.4
django-mptt-admin==0.2.1
django-nose==1.4.3
django-reversion<=1.7.99
django-uuslug

fabric
johnny-cache
pillow
psycopg2
python-memcached
south<2.0
pbkdf2
